Create AVL_Tree
Basic implementation of a simple AVL Tree data structure. A lot of "Standard" functions and overloads are missing. 
Iterators are missing on purpose as well: this implementation is supposed to be read by beginners unlike the previous ones, and also, this is only meant for learning purposes. 
The implementation only includes basic functionality, nothing too fancy here.